A Grand Update to 1.13!
So, as it turns out, I could have used an event this whole time. Shh...don't tell Lex; he'll kill me.
In my defense, when I first wrote this, I was aware that Myrathi's version was a coremod, so I thought it was the only way to do it. Plus, I was a dumb kid. Now I'm a dumb adult. Big difference.
As a nice side effect, this mod should work with modded bows too now! (As long as the modded bow fires the event I need.)

I feel too lazy to backport this to 1.12 however. Let it be a reminder of mistakes made in the past.