Update to MC 1.20.5/6.
Mojang completely refactored the Bow code and pulled most of the logic into the superclass, which Crossbows also inherit from. Crossbows now support infinity as well. Luckily Forge already had an event for this that was equivalent for what I did in Fabric, so I just switched to using it to apply my "fix". Also needed to double my gametests to handle crossbows (and rename some stuff). Should this have always been the way? idk, but it works now!