Update to MC 1.21.
Mojang has done it again! Enchantment rework made me jostle around in my head about how to do this now. It was between literally checking the Infinity enchantment on the bow/crossbow just as before, or running the `processAmmoUse()` function to see if _any_ of the enchantments on the bow/crossbow reduced the ammo use count to 0. I went with the latter because it seems to offer better mod/datapack compatibility.

Also use new(?) Neo functions for checking if an arrow is infinite, and getting the "default creative arrow" to use for the Infinity arrow.

GameTests were broken for some reason and I can't be troubled to figure out why. Manually tested, it works.