Update to MC 1.21
Mojang has done it again! Enchantment rework made me jostle around in my head about how to do this now. It was between literally checking the Infinity enchantment on the bow/crossbow just as before, or running the `processAmmoUse()` function to see if _any_ of the enchantments on the bow/crossbow reduced the ammo use count to 0. I went with the latter because it seems to offer better mod/datapack compatibility.