Fix #11.
With the new way I implemented this mod for 1.21, the ClientLevel would think the bow failed, but the ServerLevel would know it was actually successful. If the user didn't have anything in their off-hand, this would work fine. However if they did have something in their off-hand, and it was a long-use item like a shield, then it would look like you are using the shield the whole time. This was a visual glitch that I have now corrected.

There is still technically a bug where if you have a quick-use item (like placing a block for example), it will still use that, and then you would use your bow. I tried for a long time to fix that, but I think there's just a fundamental problem in Minecraft logic on that. So I'm just living with that shortcoming for now.

Also, should this be upstream? Maybe...